api/requirements.txt
# FastAPI and web server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0

# AI/ML
torch>=2.0.0
transformers>=4.30.0
torchvision>=0.15.0

# Database
supabase>=2.0.0
postgrest>=0.11.0
httpx>=0.24.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0